refactor(DAO-2005): migrate btc-vault formatters to shared utilities
- Remove formatPercent, date functions from btc-vault formatters.ts
- Update mappers.ts to import formatPercentage, formatDateDayFirst,
  formatDateMonthFirst from @/lib/utils
- Update BtcVaultMetrics to use formatDateFullMonth and formatDateMonthFirst
  from @/lib/utils
