+++ src/app/btc-vault/services/ui/mappers.ts
@@ -4,7 +4,13 @@ import { DEPOSIT_ACTIONS } from '@/app/api/btc-vault/v1/schemas'
 import { formatSymbol, getFiatAmount } from '@/app/shared/formatter'
 import Big from '@/lib/big'
 import { RBTC } from '@/lib/constants'
-import { formatCurrencyWithLabel, shortAddress } from '@/lib/utils'
+import {
+  formatCurrencyWithLabel,
+  formatDateDayFirst,
+  formatDateMonthFirst,
+  formatPercentage,
+  shortAddress,
+} from '@/lib/utils'
 
 import {
   ACTIVE_REQUEST_REASON,
@@ -30,15 +36,7 @@ import type {
 } from '../types'
 import { lockedSharePriceFromEpochSnapshot, lockedSharePriceToNavPerHumanShareWei } from '../vaultShareNav'
 import type { BtcVaultHistoryApiResponse, BtcVaultHistoryItemWithStatus } from './api-types'
-import {
-  formatApyPercent,
-  formatCountdown,
-  formatDateMonthFirst,
-  formatDateShort,
-  formatPercent,
-  formatTimestamp,
-  shortenTxHash,
-} from './formatters'
+import { formatApyPercent, formatCountdown, shortenTxHash } from './formatters'
 import type {
   ActionEligibility,
   ActiveRequestDisplay,
@@ -128,7 +126,7 @@ export function toEpochDisplay(raw: EpochState): EpochDisplay {
     raw.status === 'open'
       ? `Closes in ${formatCountdown(raw.endTime)}`
       : raw.status === 'claimable' && raw.settledAt != null
-        ? `Settled ${formatTimestamp(raw.settledAt)}`
+        ? `Settled ${formatDateDayFirst(raw.settledAt)}`
         : raw.status.charAt(0).toUpperCase() + raw.status.slice(1)
   return {
     epochId: raw.epochId,
@@ -136,7 +134,7 @@ export function toEpochDisplay(raw: EpochState): EpochDisplay {
     statusSummary,
     isAcceptingRequests,
     endTime: raw.endTime,
-    closesAtFormatted: formatDateShort(raw.endTime),
+    closesAtFormatted: formatDateDayFirst(raw.endTime),
   }
 }
 
@@ -162,7 +160,7 @@ export function toUserPositionDisplay(raw: UserPosition, rbtcPrice: number): Use
     rbtcBalanceFormatted: formatSymbol(raw.rbtcBalance, RBTC),
     vaultTokensFormatted: formatSymbol(raw.vaultTokens, 'ctokenvault'),
     positionValueFormatted: formatSymbol(raw.positionValue, RBTC),
-    percentOfVaultFormatted: formatPercent(raw.percentOfVault),
+    percentOfVaultFormatted: formatPercentage(raw.percentOfVault),
     vaultTokensRaw: raw.vaultTokens,
     rbtcBalanceRaw: raw.rbtcBalance,
 
@@ -171,7 +169,7 @@ export function toUserPositionDisplay(raw: UserPosition, rbtcPrice: number): Use
     currentEarningsFormatted: formatSymbol(currentEarnings, RBTC),
     totalBalanceFormatted: formatSymbol(raw.positionValue, RBTC),
     totalBalanceRaw: raw.positionValue,
-    yieldPercentToDateFormatted: formatPercent(yieldPercent),
+    yieldPercentToDateFormatted: formatPercentage(yieldPercent),
 
     fiatWalletBalance: hasFiat ? formatCurrencyWithLabel(getFiatAmount(raw.rbtcBalance, rbtcPrice)) : null,
     fiatVaultShares: hasFiat ? formatCurrencyWithLabel(getFiatAmount(raw.positionValue, rbtcPrice)) : null,
@@ -325,7 +323,7 @@ export function toActiveRequestDisplay(
     type: req.type,
     amountFormatted,
     status: req.status,
-    createdAtFormatted: formatTimestamp(req.timestamps.created),
+    createdAtFormatted: formatDateDayFirst(req.timestamps.created),
     claimable: claimableInfo?.claimable ?? false,
     lockedSharePriceFormatted:
       claimableInfo?.lockedSharePrice != null
@@ -334,7 +332,7 @@ export function toActiveRequestDisplay(
     finalizeId: req.type === 'deposit' ? req.epochId : req.batchRedeemId,
     epochId: req.epochId,
     batchRedeemId: req.batchRedeemId,
-    lastUpdatedFormatted: formatDateShort(lastUpdated),
+    lastUpdatedFormatted: formatDateDayFirst(lastUpdated),
     sharesFormatted,
     usdEquivalentFormatted,
     ...(req.displayStatus && { displayStatus: req.displayStatus }),
@@ -425,8 +423,8 @@ export function toPaginatedHistoryDisplay(
         type: req.type,
         amountFormatted,
         status: req.status,
-        createdAtFormatted: formatTimestamp(req.timestamps.created),
-        finalizedAtFormatted: req.timestamps.finalized ? formatTimestamp(req.timestamps.finalized) : null,
+        createdAtFormatted: formatDateDayFirst(req.timestamps.created),
+        finalizedAtFormatted: req.timestamps.finalized ? formatDateDayFirst(req.timestamps.finalized) : null,
         submitTxShort: req.txHashes.submit ? shortenTxHash(req.txHashes.submit) : null,
         finalizeTxShort: req.txHashes.finalize ? shortenTxHash(req.txHashes.finalize) : null,
         submitTxFull: req.txHashes.submit ?? null,
@@ -435,7 +433,7 @@ export function toPaginatedHistoryDisplay(
         displayStatusLabel,
         fiatAmountFormatted,
         claimTokenType: isDeposit ? ('rbtc' as const) : ('shares' as const),
-        updatedAtFormatted: formatDateShort(updatedAt),
+        updatedAtFormatted: formatDateDayFirst(updatedAt),
       }
     }),
     total: raw.total,
@@ -546,7 +544,7 @@ export function apiHistoryToPaginatedDisplay(
           ? formatCurrencyWithLabel(Big(formatEther(amountWei)).mul(rbtcPrice))
           : null,
       claimTokenType: isDeposit ? ('rbtc' as const) : ('shares' as const),
-      updatedAtFormatted: formatDateShort(item.timestamp),
+      updatedAtFormatted: formatDateDayFirst(item.timestamp),
     }
   })
